resolved PR comments and done with destructuring
