+++ app/javascript/src/components/Team/Details/PersonalDetails/Edit/MobileEditPage.tsx
@@ -129,7 +129,9 @@ const MobileEditDetails = ({
               type="text"
               value={
                 personalDetails.date_of_birth &&
-                dayjs(personalDetails.date_of_birth).format("DD.MM.YYYY")
+                dayjs(personalDetails.date_of_birth).format(
+                  personalDetails.date_format
+                )
               }
               onChange={e => {
                 updateBasicDetails(e.target.value, "date_of_birth", false);
